% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Bessel.R
\name{Bessel}
\alias{Bessel}
\title{Type one Bessel function of Herz}
\usage{
Bessel(m, x, gamma)
}
\arguments{
\item{m}{truncation weight, a positive integer}

\item{x}{either a real symmetric matrix, a Hermitian complex matrix,
or a numeric or complex vector, the eigen values of the matrix}

\item{gamma}{the gamma parameter, real or complex number with \code{Re(gamma)>-1}}
}
\value{
A real or complex number.
}
\description{
Evaluates the type one Bessel of Herz.
}
\examples{
# for a single x, the relation with the Bessel J-function:
t <- 2
gamma <- 3
besselJ(t, gamma)
Bessel(m=15, t^2/4, gamma) * (t/2)^gamma
# it also holds for a complex variable:
t <- 1 + 2i
Bessel::BesselJ(t, gamma)
Bessel(m=15, t^2/4, gamma) * (t/2)^gamma
}
\references{
A. K. Gupta and D. K. Nagar.
\emph{Matrix variate distributions}. Chapman and Hall, 1999.
}
