% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{hypergeomPFQ_julia}
\alias{hypergeomPFQ_julia}
\title{Evaluation with Julia}
\usage{
hypergeomPFQ_julia()
}
\value{
A function with the same arguments as \code{\link{hypergeomPFQ}}.
}
\description{
Evaluate the hypergeometric function of a matrix argument with
  Julia. This is highly faster.
}
\note{
See \code{\link[JuliaConnectoR]{JuliaConnectoR-package}} for
  information about setting up Julia. If you want to directly use Julia,
  you can use \href{https://github.com/stla/HypergeoMat.jl}{my package}.
}
\examples{
library(HypergeoMat)
\donttest{if(JuliaConnectoR::juliaSetupOk()){
  jhpq <- hypergeomPFQ_julia()
  jhpq(30, c(1+1i, 2, 3), c(4, 5), c(0.1, 0.2, 0.3+0.3i))
  JuliaConnectoR::stopJulia()
}}
}
