% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiGammaBeta.R
\name{mvgamma}
\alias{mvgamma}
\alias{lmvgamma}
\title{Multivariate Gamma function (of complex variable)}
\usage{
lmvgamma(x, p)

mvgamma(x, p)
}
\arguments{
\item{x}{a real or a complex number; \code{Re(x)>0} for \code{lmvgamma} and
\code{x} must not be a negative integer for \code{mvgamma}}

\item{p}{a positive integer, the dimension}
}
\value{
A real or a complex number.
}
\description{
The multivariate Gamma function (\code{mvgamma}) and
its logarithm (\code{lmvgamma}).
}
\examples{
x <- 5
mvgamma(x, p = 2)
sqrt(pi)*gamma(x)*gamma(x-1/2)
}
