% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/julia.R
\name{hypergeomPFQ_julia}
\alias{hypergeomPFQ_julia}
\title{Evaluation with Julia}
\usage{
hypergeomPFQ_julia()
}
\value{
A function with the same arguments as \code{\link{hypergeomPFQ}}.
}
\description{
Evaluate the hypergeometric function of a matrix argument with
Julia. This is highly faster.
}
